+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/api/dto/filter/FormableFilter.java
@@ -4,14 +4,29 @@ package eu.bcvsolutions.idm.core.api.dto.filter;
  * Filter for form-able entities
  *
  * @author svandav
+ * @author Radek Tomiška
  */
-public interface FormableFilter extends BaseFilter {
+public interface FormableFilter extends BaseDataFilter {
 
 	// If is true, then will be to the result added EAV metadata (it means FormInstances) 
 	String PARAMETER_ADD_EAV_METADATA = "addEavMetadata";
 	
-    Boolean getAddEavMetadata();
+	/**
+	 * Load extended attributes after filter is applied. 
+	 * 
+	 * @return true - extended attributes will be loaded
+	 */
+    default Boolean getAddEavMetadata() {
+    	return getParameterConverter().toBoolean(getData(), PARAMETER_ADD_EAV_METADATA);
+    }
 
-    void setAddEavMetadata(Boolean property);
+    /**
+     * Load extended attributes after filter is applied.
+     * 
+     * @param value true - extended attributes will be loaded
+     */
+    default void setAddEavMetadata(Boolean value) {
+    	set(PARAMETER_ADD_EAV_METADATA, value);
+    }
 
 }