+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/api/dto/filter/IdmIncompatibleRoleFilter.java
@@ -5,7 +5,6 @@ import java.util.UUID;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 
-import eu.bcvsolutions.idm.core.api.domain.ExternalIdentifiable;
 import eu.bcvsolutions.idm.core.api.domain.RequestFilterPredicate;
 import eu.bcvsolutions.idm.core.api.dto.IdmIncompatibleRoleDto;
 import eu.bcvsolutions.idm.core.api.utils.DtoUtils;
@@ -17,7 +16,7 @@ import eu.bcvsolutions.idm.core.api.utils.DtoUtils;
  * @author Radek Tomiška 
  * @since 9.4.0
  */
-public class IdmIncompatibleRoleFilter extends DataFilter implements ExternalIdentifiable {
+public class IdmIncompatibleRoleFilter extends DataFilter implements ExternalIdentifiableFilter {
 	/**
 	 * Superior role
 	 */
@@ -51,15 +50,5 @@ public class IdmIncompatibleRoleFilter extends DataFilter implements ExternalIde
 	
 	public void setSubId(UUID subId) {
 		data.set(PARAMETER_SUB_ID, subId);
-	}
-	
-	@Override
-	public String getExternalId() {
-		return (String) data.getFirst(PROPERTY_EXTERNAL_ID);
-	}
-	
-	@Override
-	public void setExternalId(String externalId) {
-		data.set(PROPERTY_EXTERNAL_ID, externalId);
 	}
 }