+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/api/dto/filter/IdmTokenFilter.java
@@ -6,7 +6,6 @@ import java.util.UUID;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 
-import eu.bcvsolutions.idm.core.api.domain.ExternalIdentifiable;
 import eu.bcvsolutions.idm.core.api.dto.IdmTokenDto;
 import eu.bcvsolutions.idm.core.api.utils.DtoUtils;
 
@@ -16,7 +15,7 @@ import eu.bcvsolutions.idm.core.api.utils.DtoUtils;
  * @author Radek Tomiška
  * @since 8.2.0
  */
-public class IdmTokenFilter extends DataFilter implements ExternalIdentifiable {
+public class IdmTokenFilter extends DataFilter implements ExternalIdentifiableFilter {
 	
 	public static final String PARAMETER_OWNER_ID = "ownerId";
 	public static final String PARAMETER_OWNER_TYPE = "ownerType";
@@ -69,15 +68,5 @@ public class IdmTokenFilter extends DataFilter implements ExternalIdentifiable {
 	
 	public void setExpirationTill(ZonedDateTime expirationTill) {
 		data.set(PARAMETER_EXPIRATION_TILL, expirationTill);
-	}
-	
-	@Override
-	public String getExternalId() {
-		return (String) data.getFirst(PROPERTY_EXTERNAL_ID);
-	}
-	
-	@Override
-	public void setExternalId(String externalId) {
-		data.set(PROPERTY_EXTERNAL_ID, externalId);
 	}
 }