+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/api/dto/filter/IdmTreeNodeFilter.java
@@ -5,7 +5,6 @@ import java.util.UUID;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 
-import eu.bcvsolutions.idm.core.api.domain.ExternalIdentifiable;
 import eu.bcvsolutions.idm.core.api.dto.IdmTreeNodeDto;
 import eu.bcvsolutions.idm.core.api.utils.DtoUtils;
 
@@ -15,7 +14,7 @@ import eu.bcvsolutions.idm.core.api.utils.DtoUtils;
  * @author Ondrej Kopr <kopr@xyxy.cz>
  * @author Radek Tomiška
  */
-public class IdmTreeNodeFilter extends DataFilter implements CorrelationFilter, ExternalIdentifiable {
+public class IdmTreeNodeFilter extends DataFilter implements CorrelationFilter, ExternalIdentifiableFilter {
 
 	public static final String PARAMETER_CODE = "code"; // PARAMETER_CODEABLE_IDENTIFIER can be used too
 	/**
@@ -56,7 +55,7 @@ public class IdmTreeNodeFilter extends DataFilter implements CorrelationFilter, 
     }
 
     public void setTreeTypeId(UUID treeTypeId) {
-    	data.set(PARAMETER_TREE_TYPE_ID, treeTypeId);
+    	set(PARAMETER_TREE_TYPE_ID, treeTypeId);
     }
     
     public UUID getParent() {
@@ -64,7 +63,7 @@ public class IdmTreeNodeFilter extends DataFilter implements CorrelationFilter, 
 	}
 
 	public void setParent(UUID parent) {
-		data.set(PARAMETER_PARENT, parent);
+		set(PARAMETER_PARENT, parent);
 	}
 
     public Boolean getDefaultTreeType() {
@@ -72,7 +71,7 @@ public class IdmTreeNodeFilter extends DataFilter implements CorrelationFilter, 
     }
 
     public void setDefaultTreeType(Boolean defaultTreeType) {
-    	data.set(PARAMETER_DEFAULT_TREE_TYPE, defaultTreeType);
+    	set(PARAMETER_DEFAULT_TREE_TYPE, defaultTreeType);
     }
 
     public boolean isRecursively() {
@@ -80,45 +79,15 @@ public class IdmTreeNodeFilter extends DataFilter implements CorrelationFilter, 
     }
 
     public void setRecursively(boolean recursively) {
-    	data.set(PARAMETER_RECURSIVELY, recursively);
+    	set(PARAMETER_RECURSIVELY, recursively);
     }
-
-	@Override
-	public String getProperty() {
-		return (String) data.getFirst(PARAMETER_CORRELATION_PROPERTY);
-	}
-
-	@Override
-	public void setProperty(String property) {
-		data.set(PARAMETER_CORRELATION_PROPERTY, property);
-	}
-
-	@Override
-	public String getValue() {
-		return (String) data.getFirst(PARAMETER_CORRELATION_VALUE);
-	}
-
-	@Override
-	public void setValue(String value) {
-		data.set(PARAMETER_CORRELATION_VALUE, value);
-	}
     
     public String getCode() {
 		return (String) data.getFirst(PARAMETER_CODE);
 	}
 
 	public void setCode(String code) {
-		data.set(PARAMETER_CODE, code);
-	}
-	
-	@Override
-	public String getExternalId() {
-		return (String) data.getFirst(PROPERTY_EXTERNAL_ID);
-	}
-	
-	@Override
-	public void setExternalId(String externalId) {
-		data.set(PROPERTY_EXTERNAL_ID, externalId);
+		set(PARAMETER_CODE, code);
 	}
 	
 	/**
@@ -134,6 +103,6 @@ public class IdmTreeNodeFilter extends DataFilter implements CorrelationFilter, 
 	 * @param roots
 	 */
 	public void setRoots(Boolean roots) {
-		data.set(PARAMETER_ROOTS, roots);
+		set(PARAMETER_ROOTS, roots);
 	}
 }