+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/api/dto/filter/IdmTreeTypeFilter.java
@@ -3,7 +3,6 @@ package eu.bcvsolutions.idm.core.api.dto.filter;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 
-import eu.bcvsolutions.idm.core.api.domain.ExternalIdentifiable;
 import eu.bcvsolutions.idm.core.api.dto.IdmTreeTypeDto;
 
 /**
@@ -11,7 +10,7 @@ import eu.bcvsolutions.idm.core.api.dto.IdmTreeTypeDto;
  *
  * @author Radek Tomiška
  */
-public class IdmTreeTypeFilter extends DataFilter implements ExternalIdentifiable {
+public class IdmTreeTypeFilter extends DataFilter implements ExternalIdentifiableFilter {
 
 	public static final String PARAMETER_CODE = "code"; // PARAMETER_CODEABLE_IDENTIFIER can be used too
 
@@ -29,15 +28,5 @@ public class IdmTreeTypeFilter extends DataFilter implements ExternalIdentifiabl
 
 	public void setCode(String code) {
 		data.set(PARAMETER_CODE, code);
-	}
-	
-	@Override
-	public String getExternalId() {
-		return (String) data.getFirst(PROPERTY_EXTERNAL_ID);
-	}
-	
-	@Override
-	public void setExternalId(String externalId) {
-		data.set(PROPERTY_EXTERNAL_ID, externalId);
 	}
 }