+++ Realization/backend/core/core-impl/src/main/java/eu/bcvsolutions/idm/core/model/repository/filter/ModifiedTillFilterBuilder.java
@@ -12,46 +12,47 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
 
 import eu.bcvsolutions.idm.core.api.dto.filter.DataFilter;
-import eu.bcvsolutions.idm.core.api.dto.filter.ModifiedFromFilter;
+import eu.bcvsolutions.idm.core.api.dto.filter.ModifiedTillFilter;
 import eu.bcvsolutions.idm.core.api.entity.AbstractEntity;
 import eu.bcvsolutions.idm.core.api.entity.AbstractEntity_;
 import eu.bcvsolutions.idm.core.api.repository.filter.BaseFilterBuilder;
 
 /**
- * Modified from filter (auditable.modified or uditable.created  >= modifiedFrom).
+ * Filter for filtering entities changed till given time stamp (auditable.modified or uditable.created <= modifiedTill).
  * Created date is used as fallback, if modified is {@code null} => creation is the last modification.
  * 
- * @author Vít Švanda
+ * @author Radek Tomiška
+ * @since 10.2.0
  */
-@Component
-public class ModifiedFromableFilter<E extends AbstractEntity> extends BaseFilterBuilder<E, DataFilter> {
+@Component()
+public class ModifiedTillFilterBuilder<E extends AbstractEntity> extends BaseFilterBuilder<E, DataFilter> {
 
+	
+	
 	@Override
 	public String getName() {
-		return ModifiedFromFilter.PARAMETER_MODIFIED_FROM;
+		return ModifiedTillFilter.PARAMETER_MODIFIED_TILL;
 	}
 
 	@Override
 	public Predicate getPredicate(Root<E> root, AbstractQuery<?> query, CriteriaBuilder builder, DataFilter filter) {
-
-		ZonedDateTime modifiedFrom = filter.getModifiedFrom();
+		ZonedDateTime modifiedTill = filter.getModifiedTill();
 		
-		if (modifiedFrom == null) {
+		if (modifiedTill == null) {
 			return null;
 		}
 
 		return builder.or(
-				builder.greaterThanOrEqualTo(root.get(AbstractEntity_.modified), modifiedFrom),
+				builder.lessThanOrEqualTo(root.get(AbstractEntity_.modified), modifiedTill),
 				builder.and(
 						builder.isNull(root.get(AbstractEntity_.modified)), // modified is null => creation is last modification
-						builder.greaterThanOrEqualTo(root.get(AbstractEntity_.created), modifiedFrom)
+						builder.lessThanOrEqualTo(root.get(AbstractEntity_.created), modifiedTill)
 				)
 		);
 	}
 
 	@Override
 	public Page<E> find(DataFilter filter, Pageable pageable) {
-		throw new UnsupportedOperationException(
-				"Find by modified from only is not supported, use concrete service instead.");
+		throw new UnsupportedOperationException("Find by modified till only is not supported, use concrete service instead.");
 	}
 }